+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Latest (add to next release)
 ------------------------------
+*March 3, 2021*
+
+### Added
+- Component tests for link button
+
 *February 25, 2021*
 
 ### Changed